requirements.txt
flask
flask-socketio
paho-mqtt
requests
eventlet
